requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-jwt-extended>=4.6.0
pymysql>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
cryptography>=42.0.0

